pyproject.toml
[tool.poetry]
name = "code-explainer"
version = "0.1.0"
description = "Explains code using AI"
authors = ["Fernando Cordeiro"]
readme = "README.md"
packages = [{include = "code_explainer"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.6"
httpx = "^0.24.0"
python-dotenv = "^1.0.0"
streamlit = "^1.22.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"
black = "^23.3.0"
ruff = "^0.0.265"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
